+++ src/main/java/com/mraof/minestuck/item/CassetteItem.java
@@ -3,24 +3,37 @@ package com.mraof.minestuck.item;
 import com.mraof.minestuck.block.CassettePlayerBlock;
 import com.mraof.minestuck.block.EnumCassetteType;
 import com.mraof.minestuck.block.MSBlocks;
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.ComponentUtils;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.network.chat.Style;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.JukeboxSong;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 
+import javax.annotation.ParametersAreNonnullByDefault;
+import java.util.List;
+
+@ParametersAreNonnullByDefault
 public class CassetteItem extends Item
 {
-	public final EnumCassetteType cassetteID;
+	public final EnumCassetteType cassetteType;
 	private final int comparatorValue;
 	
-	public CassetteItem(int comparatorValue, EnumCassetteType cassetteName, Properties builder, int lengthInTicks)
+	public CassetteItem(int comparatorValue, EnumCassetteType cassetteType, Properties builder, int lengthInTicks)
 	{
 		super(builder);
 		this.comparatorValue = comparatorValue;
-		this.cassetteID = cassetteName;
+		this.cassetteType = cassetteType;
 	}
 	
 	@Override
@@ -42,6 +55,23 @@ public class CassetteItem extends Item
 		} else
 		{
 			return InteractionResult.PASS;
+		}
+	}
+	
+	@Override
+	public void appendHoverText(ItemStack stack, TooltipContext context, List<Component> tooltipComponents, TooltipFlag tooltipFlag)
+	{
+		ResourceKey<JukeboxSong> songKey = this.cassetteType.getJukeboxSong();
+		HolderLookup.Provider registries = context.registries();
+		if (songKey != null && registries != null)
+		{
+			registries.lookup(songKey.registryKey())
+					.flatMap(registry -> registry.get(songKey))
+					.ifPresent(song -> {
+				MutableComponent songDescription = song.value().description().copy();
+				ComponentUtils.mergeStyles(songDescription, Style.EMPTY.withColor(ChatFormatting.GRAY));
+				tooltipComponents.add(songDescription);
+			});
 		}
 	}