+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Fix dance_stab_dance and emissary_of_dance being labelled as each other
 - Fix cassette playing weapons not having a menu for storing cassette
 - Fix weapons not being enchantable
+- Add back song description for cassette items
 
 ### Removed